# docs/swagger.yaml
definitions:
...
  dto.ResImportUserExcel:
    properties:
      error_message:
        description: Message error jika status failed
        type: string
      row:
        description: Nomor baris di Excel
        type: integer
      status:
        description: 'Status row: "success" atau "failed"'
        type: string
      username:
        description: Username user
        type: string
    type: object
  dto.ResImportUsers:
    properties:
      failed_count:
        type: integer
      results:
        items:
          $ref: '#/definitions/dto.ResImportUserExcel'
        type: array
      success_count:
        type: integer
      total_rows:
        type: integer
    type: object
...
  response.NonPaginationResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
...
info:
  contact: {}
  description: Welcome to the API documentation for the Base Template Web Application.
    This comprehensive guide is designed to help developers seamlessly integrate and
    interact with our platform's functionalities. Whether you're building new features,
    enhancing existing ones, or troubleshooting, this documentation provides all the
    necessary resources and information.
  title: Base Template API Documentation
  version: 0.0-beta
paths:
...
  /v1/user-management/user/import:
    post:
      consumes:
      - multipart/form-data
      description: 'Import multiple users from an Excel file (.xlsx or .xls). The
        Excel file must have columns: email, full_name, username, nik, role_name.
        Validates for duplicate email, username, and NIK. Returns HTTP 400 if any
        row fails with detailed error information per row.'
      parameters:
      - description: 'Excel file (.xlsx or .xls) with columns: email, full_name, username,
          nik, role_name'
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Successfully imported all users
          schema:
            allOf:
            - $ref: '#/definitions/response.NonPaginationResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ResImportUsers'
              type: object
        "400":
          description: Bad request - one or more rows failed validation. Response
            contains details for each row including row number, username, status,
            and error message
          schema:
            allOf:
            - $ref: '#/definitions/response.NonPaginationResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ResImportUsers'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
      security:
      - BearerAuth: []
      summary: Import users from Excel file
      tags:
      - User Management
  /v1/user-management/user/import/template:
    get:
      consumes:
      - application/json
      description: 'Download Excel template file for importing users. Template contains
        columns: email, full_name, username, nik, role_name with example data.'
      produces:
      - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
      responses:
        "200":
          description: Excel template file
          schema:
            type: file
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
      security:
      - BearerAuth: []
      summary: Download user import Excel template
      tags:
      - User Management
...
securityDefinitions:
  BearerAuth:
    description: 'Enter JWT token (ex: Bearer eyJhbGciOiJIU....)'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"